agregando metodos de clase jugador

+++ src/main/java/com/futboldatos/jugadores/service/EquipoService.java
@@ -4,6 +4,7 @@ package com.futboldatos.jugadores.service;
 import com.futboldatos.jugadores.model.Equipo;
 import com.futboldatos.jugadores.repository.IEquipoRepository;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -27,9 +28,16 @@ public class EquipoService implements IEquipoService{
     }
 
     @Override
-    public void borrarEquipo(Long id) {
+    public boolean borrarEquipo(Long id) {
+    Optional<Equipo> equipo = equipoRepo.findById(id);
+
+    if (equipo.isPresent()) {
         equipoRepo.deleteById(id);
+        return true; // Indica que se eliminó el equipo con éxito
+    } else {
+        return false; // Indica que no se encontró un equipo con el ID proporcionado
     }
+}
 
     @Override
     public Equipo buscarEquipo(Long id) {
@@ -38,8 +46,14 @@ public class EquipoService implements IEquipoService{
     }
 
     @Override
-    public void editarEquipo(Equipo equipo) {
-        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
+    public void editarEquipo(Long idOriginal, String nombreNuevo, String paisNuevo, List jugadoresNuevos) {
+        
+        Equipo equipo = this.buscarEquipo(idOriginal);
+        equipo.setNombre(nombreNuevo);
+        equipo.setPais(paisNuevo);
+        equipo.setListaJugadores(jugadoresNuevos);
+        
+        this.crearEquipo(equipo);
     }
     
 }
